fix(ken-551): a drift detail is a path, not a rendering [no-changelog]
unmanaged() and unmanaged_compared() still escaped the path into
DriftRow::detail, so also_at deduplicated renderings while its comment
claimed it deduplicated places, and conflict_detail escaped a second time
on top. The path is stored as it is now, and every reader escapes exactly
once at its own print: conflict_detail, also_at, print_unmanaged and
verify. takeover's note and adopt's error fields keep escaping where they
are built, being finished sentences their readers print as they stand.

Tested through the production constructors rather than a hand-built row,
which is what the last control missed: a project directory carrying a real
control character reaches the row as the path it is, and reaches the
terminal as its own characters. Two positions that render alike are not
constructible through the CLI — a position's variable part is the project
root, shared by every tool the item is blocked for — so the collapse
itself stays covered by the unit control on also_at.

The test file crossed its 800-line cap and splits at the seam between what
the run reports at the end and what one conflict says.
